+++ src/lib/types/proposal.ts
@@ -11,6 +11,7 @@ export enum ProposalStatus {
   REJECTED = "Rejected",
   FAILED = "Failed",
   DEPOSIT_FAILED = "Inactive",
+  CANCELLED = "Cancelled",
 }
 export const zProposalStatus = z.nativeEnum(ProposalStatus);
 
@@ -95,12 +96,18 @@ export interface ProposalDeposit {
   txHash: string;
 }
 
+// TODO: combine with MsgBody in services/tx.ts
+interface Message {
+  "@type": string;
+  [key: string]: unknown;
+}
+
 export interface ProposalData extends Proposal {
   createdHeight: Nullable<number>;
   createdTimestamp: Nullable<Date>;
   createdTxHash: Nullable<string>;
   description: string;
-  messages: Nullable<unknown[]>;
+  messages: Nullable<Message[]>;
   metadata: string;
   proposalDeposits: ProposalDeposit[];
   resolvedTimestamp: Nullable<Date>;